Changed existing definitions to allow credentials having int64 values rather than int32 - I'm assuming our system will run on x86_64 machines

+++ common/comm/commands/commands.go
@@ -344,7 +344,7 @@ func NewGetCredentialRequest(lambda *coconut.Lambda,
 func NewSpendCredentialRequest(sig *coconut.Signature,
 	pubM []*Curve.BIG,
 	theta *coconut.ThetaTumbler,
-	val int32,
+	val int64,
 	address []byte,
 ) (*SpendCredentialRequest, error) {
 	protoSig, err := sig.ToProto()
